Use size_t when strings are involved.
Use strtoul to parse unsigned ints.

+++ util.h
@@ -7,10 +7,10 @@ extern "C" {
 int32_t str_change_chr(char *str, char from, char to);
 int32_t str_trim_end(char *str, char c);
 const char *str_find_first_char(const char *str, const char *chars);
-const char *str_find_len_first_char(const char *str, int32_t str_len, const char *chars);
-const char *str_find_len_char(const char *str, int32_t str_len, char c);
-const char *str_find_len_str(const char *str, int32_t str_len, const char *find);
-const char *str_find_len_case_str(const char *str, int32_t str_len, const char *find);
+const char *str_find_len_first_char(const char *str, size_t str_len, const char *chars);
+const char *str_find_len_char(const char *str, size_t str_len, char c);
+const char *str_find_len_str(const char *str, size_t str_len, const char *find);
+const char *str_find_len_case_str(const char *str, size_t str_len, const char *find);
 
 char *dns_resolve(const char *host, int32_t *error);
 char *get_url_host(const char *url);
